Refactor onboarding and clearml utils to prevent errors on aqua
